добавляет модули util.js и data.js
